funcionalidad policie role,permission, y falta terminar usuario

+++ app/Http/Controllers/PermissionController.php
@@ -10,6 +10,9 @@ use Illuminate\Http\Request;
 
 class PermissionController extends Controller
 {
+    public function __construct(){
+        $this->middleware('role:'.config('app.admin_role'));
+    }
     /**
      * Display a listing of the resource.
      *
@@ -18,6 +21,7 @@ class PermissionController extends Controller
     public function index()
     {
         //
+        $this->authorize('index',Permission::class);
         $permissions=Permission::all();
         return view('theme.backoffice.pages.permission.index',['permissions'=>$permissions]);
     }
@@ -30,6 +34,7 @@ class PermissionController extends Controller
     public function create()
     {
         //
+        $this->authorize('create',Permission::class);
         $roles=Role::all();
         return view('theme.backoffice.pages.permission.create',['roles'=>$roles]);
 
@@ -57,6 +62,7 @@ class PermissionController extends Controller
     public function show(Permission $permission)
     {
         //
+        $this->authorize('view',$permission);
         return view('theme.backoffice.pages.permission.show',['permission'=>$permission]);
     }
 
@@ -69,6 +75,7 @@ class PermissionController extends Controller
     public function edit(Permission $permission)
     {
         //
+        $this->authorize('update',$permission);
         $roles=Role::all();
 
         return view('theme.backoffice.pages.permission.edit',['permission'=>$permission,'roles'=>$roles]);
@@ -97,6 +104,7 @@ class PermissionController extends Controller
     public function destroy(Permission $permission)
     {
         //
+        $this->authorize('delete',$permission);
         $role=$permission->role;
         $permission->delete();
         alert('Exito','Permiso eliminado','success');
